Fixed return type hint in unions.md
I know that it seems a small, nearly stylistic, change but touches an important aspect of what get_discriminator value function should actually return if input data is unspecified. 

It can be argued that str should be returned, as a type hint suggest, and a ValueError or other Exception should be raised when the input doesn't match expected structure.

Another option would be to return Any - the same way before validation handles data: Any input if it has unexpected structer. However Discriminator documentation stands in contrast to that with discriminator: str | Callable[[Any], Hashable]

Due to uncertainty just corrected type hint of the current function but would be happy to hear which appoach is expected